requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
h11==0.13.0
httptools==0.4.0
idna==3.3
mysqlclient==2.1.0
persiantools==3.0.1
pkg_resources==0.0.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
whitenoise==6.2.0
